치환 및 quote 제거 함수 수정 중 내 대갈통 x터진다
