+++ parse/parse.c
@@ -11,285 +11,6 @@
 
 t_info g_info;
 
-int is_space(char c)
-{
-	if (c == ' ' || c == '\t' || c == '\v' || \
-			c == '\n' || c == '\f' || c == '\r')
-		return (1);
-	return (0);
-}
-
-t_token	*new_token(int type, char *word)
-{
-	t_token *node;
-
-	node = malloc(sizeof(t_token) * 1);
-	if (!node)
-		exit(1);
-	node->type = type;
-	node->word = word;
-	node->next = NULL;
-	return (node);
-}
-
-int quote_size(const char *str, char quote)
-{
-	int i;
-
-	i = 1;
-	while (str[i])
-	{
-		if (str[i] == '\0')
-			// 안 닫혔을 때: syntax error?
-			exit(1);
-		if (str[i] == quote)
-		{
-			i++;
-			break ;
-		}
-		i++;
-	}
-	return (i);
-}
-
-int rdir_size(const char *str, char rdir)
-{
-	int i;
-
-	i = 1;
-	if (str[i] == rdir)
-		i++;
-	return (i);
-}
-
-int count_chunk_size(const char *str, char sign)
-{
-	int size;
-
-	size = 0;
-	if (sign == '\'' || sign == '\"')
-		size = quote_size(str, sign);
-	else if (sign == '>' || sign == '<')
-		size = rdir_size(str, sign);
-	return (size);
-}
-
-void	add_node(int type, char *word, t_token *head)
-{
-	if (head->word == NULL)
-	{
-		head->type = type;
-		head->word = word;
-	}
-	else
-	{
-		while (head->next != NULL)
-			head = head->next;
-		head->next = new_token(type, word);
-	}
-}
-
-void	make_chunk(const char *chunk, int chunk_size, t_token *chunks)
-{
-	int		type;
-	char	*word;
-	int 	i;
-
-	if (*chunk == '\'')
-		type = S_QUOTE;
-	else if (*chunk == '\"')
-		type = D_QUOTE;
-	else if (*chunk == '>' || *chunk == '<')
-		type = REDIRECT;
-	else if (*chunk == '|')
-		type = PIPE;
-	else if (*chunk == '$')
-		type = DOLLAR;
-	else
-		type = CHAR;
-	word = malloc(sizeof(char) * (chunk_size + 1));
-	if (!word)
-		return ;
-	i = -1;
-	while (++i < chunk_size)
-		word[i] = chunk[i];
-	word[i] = '\0';
-//	ft_strlcpy(word, chunk, chunk_size);
-	add_node(type, word, chunks);
-}
-
-void	make_chunk_lst(char *input, t_token *chunks)
-{
-	char *chunk_str;
-	int chunk_size;
-
-	while (*input)
-	{
-		while (is_space(*input))
-			input++;
-		chunk_str = input;
-		chunk_size = 0;
-		if (*input == '\'' || *input == '\"' || *input == '>' || *input == '<')
-			chunk_size = count_chunk_size(input, *input);
-		else
-		{
-			while (input[chunk_size])
-			{
-				if (is_space(input[chunk_size]))
-					break ;
-				chunk_size++;
-			}
-		}
-		make_chunk(chunk_str, chunk_size, chunks);
-		input = input + chunk_size;
-	}
-}
-
-int	ft_strcmp(const char *str1, const char *str2)
-{
-	while (*str1 || *str2)
-	{
-		if (*str1 != *str2)
-			return (*str1 - *str2);
-		str1++;
-		str2++;
-	}
-	return (*str1 - *str2);
-}
-
-t_env *find_env(char *key)
-{
-	t_env *temp;
-
-	temp = g_info.env_lst;
-	while (temp && ft_strcmp(temp->key, key))
-		temp = temp->next;
-	if (!temp)
-		return (0);
-	return (temp);
-}
-
-int count_value_size(char *str)
-{
-	t_env *env;
-	char *key;
-	int word_len;
-	int key_len;
-
-	word_len = 0;
-	key_len = 0;
-	if ('a' < *str && *str < 'z')
-		while (!is_space(*(str + key_len)))
-			key_len++;
-	else if ('0' < *str && *str < '9')
-		key_len++;
-	key = malloc(sizeof(char) * (key_len + 1));
-	if (!key)
-		exit (1);
-	ft_strlcpy(key, str, key_len);
-	env = find_env(key);
-	if (env)
-		word_len += ft_strlen(env->value);
-	free(key);
-	return (word_len);
-}
-
-/* $ 하나가 아니라 두개도 세개도 올 수 있으니까 문자열 다 돌면서 확인하기
- * 1. 바로 뒤에 문자가 오면 whitespace 전까지의 문자/문자열 모두를 가져가서 일치하는 환경변수가 있는지 찾음
- * 2. 바로 뒤에 숫자가 오면 그 숫자 하나만 가져가서 일치하는 환경변수가 있는지 찾음
- * 3. 바로 뒤에 ?가 오면 마지막 에러넘을 가져와서 $? 전체를 치환
- * 4. 바로 뒤에 whitespace나 공백이 오면 $포함하여 공백 전까지 출력
- */
-
-/* 환경변수 치환 시(1, 2) split이랑 join 필요
- * split할 지점 찾기 -> split 하기 -> 환경변수 치환하기 -> join 하기
- */
-
-void ft_ptr_split(char *str, char *point, char **bowl)
-{
-
-}
-
-char	*change_word(t_token *chunk, int split_cnt)
-{
-	char **splited_words;
-	char *word;
-	char *start;
-	int word_len;
-
-	word = chunk->word;
-	word_len = 0;
-
-			// dollar가 포함 되어있을 때
-			splited_words = malloc(sizeof(char *) * (split_cnt + 1));
-			if (!splited_words)
-				exit(1);
-			ft_ptr_split(chunk->word, word, splited_words);
-			while (*splited_words)
-			{
-				if (*splited_words[0] == '$')
-					subtitute_dollar();
-				(*splited_words)++;
-			}
-
-
-//		----------------------------------------------------------------------
-			start = word + 1;
-			if ('a' < *start && *start < 'z' || \
-			'0' < *start && *start < '9')// $ 뒤에 문자나 숫자 있을 때
-				word_len += count_value_size(start);
-			else if (*start == '?') //?일 때
-				if (g_info.last_error)
-					word_len += ft_strlen(ft_itoa(g_info.last_error));
-				//last_error가 없으면(g_info 초기화시 NONE enum 필요할 듯) word_len에는 영향 x
-			else // !start || is_space(*start) || 그 외 기호들
-			{
-				word++;
-				continue;
-			}
-			word += word_len;
-//		----------------------------------------------------------------------
-}
-
-void	repair_chunk_lst(t_token *chunks)
-{
-	/* 치환 작업 및 제거
-	 * $와 "$" 뒤 문자와 일치하는 환경변수 key가 있으면 value로, 없으면 빈 문자열로 치환하면서
-	 	* $의 type은 DOLLAR -> CHAR "$"의 type은 D_QUOTE -> CHAR
-	 * ''는 ''만 제거
-	 	* ''의 type은 S_QUOTE -> CHAR
-	*/
-	t_token	*cur;
-	char *word;
-	int dollar_cnt;
-
-	cur = chunks;
-	while (cur != NULL)
-	{
-		if ((cur->type == DOLLAR) || \
-		(cur->type == D_QUOTE && ft_strchr(cur->word, '$')))
-		{
-			dollar_cnt = 0;
-			while (*(cur->word))
-			{
-				if (*(cur->word) == '$')
-					dollar_cnt++;
-				cur->word++;
-			}
-			word = change_word(cur, dollar_cnt + 1);
-		}
-		else if (cur->type == S_QUOTE)
-			word = cur->word++;
-		else
-		{
-			cur = cur->next;
-			continue ;
-		}
-		cur->word = word;
-		cur->type = CHAR;
-		cur = cur->next;
-	}
-}
 
 void	input_tokenize(char *input, t_token *tokens)
 {
@@ -307,7 +28,7 @@ void	input_tokenize(char *input, t_token *tokens)
 	make_chunk_lst(input, chunks);
 
 	/* "$" 치환 맟 '' 제거 */
-	repair_chunk_lst(chunks);
+	modify_chunk_lst(chunks);
 
 	/* while
 	 * token_lst 만들기
@@ -332,6 +53,7 @@ int	main(void)
 	}
 
 	/* 확인용 */
+
 	while (tokens != NULL)
 	{
 		printf("%s\n", tokens->word);
