+++ parse/parse.h
@@ -31,4 +31,16 @@ typedef struct s_token
 	struct s_token *next;
 }	t_token;
 
+/* utils.c */
+int is_space(char c);
+int	ft_strcmp(const char *str1, const char *str2);
+t_token	*new_token(int type, char *word);
+void	add_node(int type, char *word, t_token *head);
+
+/* chunk.c */
+void	make_chunk_lst(char *input, t_token *chunks);
+
+/* repair.c */
+void	modify_chunk_lst(t_token *chunks);
+
 #endif
